split TopNav into DesktopNav and MobileNav
